requirements.txt
matplotlib
pandas
mysql-connector-python
numpy
sqlalchemy
reportlab
openpyxl
scikit-sklearn
load_dotenv